+++ crates/holochain_cascade/src/test_utils/activity_test_data.rs
@@ -5,24 +5,24 @@ use holo_hash::EntryHash;
 use holochain_types::activity::ChainItems;
 use holochain_types::dht_op::ChainOp;
 use holochain_types::dht_op::ChainOpHashed;
-
+use holochain_types::prelude::NewEntryAction;
 use holochain_zome_types::prelude::*;
 
 /// A collection of fixtures used to create scenarios for testing the Cascade
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct ActivityTestData {
     /// AgentActivity ops to expect being able to get
-    pub hash_ops: Vec<ChainOpHashed>,
+    pub agent_activity_ops: Vec<ChainOpHashed>,
     /// "Noise", to ensure that the query filter is doing its job
-    pub noise_ops: Vec<ChainOpHashed>,
+    pub noise_agent_activity_ops: Vec<ChainOpHashed>,
     /// StoreRecord ops to expect being able to get
-    pub store_ops: Vec<ChainOpHashed>,
+    pub store_entry_ops: Vec<ChainOpHashed>,
     /// The author of the chain
     pub agent: AgentPubKey,
     /// The expected hash return values
-    pub valid_hashes: ChainItems<ActionHash>,
+    pub valid_hashes: ChainItems,
     /// The expected record return values
-    pub valid_records: ChainItems<Record>,
+    pub valid_records: ChainItems,
     /// The head of the chain produced
     pub chain_head: ChainHead,
     /// Same as the chain_head
@@ -31,7 +31,7 @@ pub struct ActivityTestData {
 
 impl ActivityTestData {
     /// Construct a set of test fixtures representing a valid source chain
-    pub fn valid_chain_scenario() -> Self {
+    pub fn valid_chain_scenario(private_entries: bool) -> Self {
         // The agent we are querying.
         let agent = fixt!(AgentPubKey);
 
@@ -39,24 +39,23 @@ impl ActivityTestData {
         let entry = Entry::App(fixt!(AppEntryBytes));
         let entry_hash = EntryHash::with_data_sync(&entry);
 
-        let to_op = |h| {
-            ChainOpHashed::from_content_sync(ChainOp::RegisterAgentActivity(fixt!(Signature), h))
-        };
+        let to_agent_activity_op =
+            |h, sig| ChainOpHashed::from_content_sync(ChainOp::RegisterAgentActivity(sig, h));
 
-        let to_record_and_op = |h: Action| {
-            let sig = fixt!(Signature);
-            // let e = Entry::App(fixt!(AppEntryBytes));
-            let op = ChainOpHashed::from_content_sync(ChainOp::StoreRecord(
+        let to_record_and_op = |a: Action, sig: Signature| {
+            let op = ChainOpHashed::from_content_sync(ChainOp::StoreEntry(
                 sig.clone(),
-                h.clone(),
-                entry.clone().into(),
+                match a {
+                    Action::Create(ref c) => NewEntryAction::Create(c.clone()),
+                    _ => unreachable!(),
+                },
+                entry.clone(),
             ));
-            let shh = SignedActionHashed::with_presigned(ActionHashed::from_content_sync(h), sig);
+            let shh = SignedActionHashed::with_presigned(ActionHashed::from_content_sync(a), sig);
             (Record::new(shh, Some(entry.clone())), op)
         };
 
-        let to_record_dna_op = |a: Action| {
-            let sig = fixt!(Signature);
+        let to_record_dna_op = |a: Action, sig: Signature| {
             let op = ChainOpHashed::from_content_sync(ChainOp::StoreRecord(
                 sig.clone(),
                 a.clone(),
@@ -74,7 +73,7 @@ impl ActivityTestData {
 
         // The store record ops for the actual data on the chain which should
         // match the set of activity ops.
-        let mut store_ops = Vec::new();
+        let mut store_entry_ops = Vec::new();
 
         // A set of activity ops:
         // - Must be on the above agents chain.
@@ -83,16 +82,17 @@ impl ActivityTestData {
         //    - Prev hash actually match prev action's hash
         //    - Seq numbers are in order.
         //    - First action must be a Dna.
-        let mut hash_ops = Vec::new();
+        let mut agent_activity_ops = Vec::new();
         let mut dna = fixt!(Dna);
         dna.author = agent.clone();
         let dna = Action::Dna(dna);
 
         // Insert the dna
-        let (el, op) = to_record_dna_op(dna.clone());
+        let dna_sig = fixt!(Signature);
+        let (el, op) = to_record_dna_op(dna.clone(), dna_sig.clone());
         valid_records.push(el);
-        store_ops.push(op);
-        hash_ops.push(to_op(dna.clone()));
+        store_entry_ops.push(op);
+        agent_activity_ops.push(to_agent_activity_op(dna.clone(), dna_sig));
 
         let creates: Vec<_> = CreateFixturator::new(Unpredictable)
             .enumerate()
@@ -106,15 +106,25 @@ impl ActivityTestData {
             create.action_seq = action_seq;
             create.prev_action = prev_hash.clone();
             create.entry_hash = entry_hash.clone();
+            create.entry_type = EntryType::App(AppEntryDef::new(
+                1.into(),
+                1.into(),
+                if private_entries {
+                    EntryVisibility::Private
+                } else {
+                    EntryVisibility::Public
+                },
+            ));
             let action = Action::Create(create);
+            let sig = fixt!(Signature);
             prev_hash = ActionHash::with_data_sync(&action);
-            hash_ops.push(to_op(action.clone()));
+            agent_activity_ops.push(to_agent_activity_op(action.clone(), sig.clone()));
 
             valid_hashes.push((action_seq, prev_hash.clone()));
 
-            let (el, op) = to_record_and_op(action);
+            let (el, op) = to_record_and_op(action, sig);
             valid_records.push(el);
-            store_ops.push(op);
+            store_entry_ops.push(op);
         }
 
         // The head of the chain is the last valid hash
@@ -131,21 +141,21 @@ impl ActivityTestData {
             hash: vec![last.1.clone()],
         };
 
-        // Finally add some random noise so we know we are getting the correct items.
-        let noise_ops = ActionFixturator::new(Unpredictable)
+        // Finally add some random noise, so we know we are getting the correct items.
+        let noise_agent_activity_ops = ActionFixturator::new(Unpredictable)
             .take(50)
-            .map(to_op)
+            .map(|a| to_agent_activity_op(a, fixt!(Signature)))
             .collect();
 
         Self {
-            hash_ops,
+            agent_activity_ops,
+            noise_agent_activity_ops,
+            store_entry_ops,
             agent,
             valid_hashes: ChainItems::Hashes(valid_hashes),
+            valid_records: ChainItems::Full(valid_records),
             highest_observed,
             chain_head,
-            noise_ops,
-            store_ops,
-            valid_records: ChainItems::Full(valid_records),
         }
     }
 }