+++ crates/holochain_integrity_types/src/action/builder.rs
@@ -1,4 +1,5 @@
 use super::EntryType;
+use super::MigrationTarget;
 use super::Timestamp;
 use crate::action;
 use crate::link::LinkTag;
@@ -13,7 +14,6 @@ use action::Dna;
 use holo_hash::ActionHash;
 use holo_hash::AgentPubKey;
 use holo_hash::AnyLinkableHash;
-use holo_hash::DnaHash;
 use holo_hash::EntryHash;
 
 #[derive(Clone, Debug)]
@@ -249,11 +249,12 @@ builder_variant!(DeleteLink {
 });
 
 builder_variant!(OpenChain {
-    prev_dna_hash: DnaHash,
+    prev_target: MigrationTarget,
+    close_hash: ActionHash,
 });
 
 builder_variant!(CloseChain {
-    new_dna_hash: DnaHash,
+    new_target: Option<MigrationTarget>,
 });
 
 builder_variant!(Create<EntryRateWeight> {
@@ -282,7 +283,7 @@ builder_variant!(AgentValidationPkg {
 /// `prev_action` field, so this helper is provided as a special case
 #[cfg(feature = "test_utils")]
 impl Dna {
-    pub fn from_builder(hash: DnaHash, builder: ActionBuilderCommon) -> Self {
+    pub fn from_builder(hash: holo_hash::DnaHash, builder: ActionBuilderCommon) -> Self {
         Self {
             author: builder.author,
             timestamp: builder.timestamp,