+++ crates/holochain_zome_types/src/fixt.rs
@@ -548,13 +548,18 @@ fixturator!(
 );
 
 fixturator!(
+    MigrationTarget;
+    variants [ Dna(DnaHash) Agent(AgentPubKey) ];
+);
+
+fixturator!(
     OpenChain;
-    constructor fn from_builder(ActionBuilderCommon, DnaHash);
+    constructor fn from_builder(ActionBuilderCommon, MigrationTarget, ActionHash);
 );
 
 fixturator!(
     CloseChain;
-    constructor fn from_builder(ActionBuilderCommon, DnaHash);
+    constructor fn from_builder(ActionBuilderCommon, MigrationTarget);
 );
 
 fixturator!(